Split hooks into individual methods and files
